BackEndAutomation/requirements.txt
allure-behave==2.8.22
allure-python-commons==2.8.22
attrs==20.3.0
behave==1.2.7.dev2
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
cucumber-tag-expressions==3.0.0
idna==2.10
importlib-metadata==3.1.0
mysql-connector-python==8.0.22
parse==1.18.0
parse-type==0.5.2
pluggy==0.13.1
protobuf==3.14.0
requests==2.25.0
six==1.15.0
urllib3==1.26.2
zipp==3.4.0
